+++ Sienar.sln
@@ -33,8 +33,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{2FD53AC3
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sienar.Utils.Tests", "shared\tests\Sienar.Utils.Tests\Sienar.Utils.Tests.csproj", "{757F285A-17D6-4387-886B-FC44445F1C1B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sienar.Utils.Blazor", "shared\src\Sienar.Utils.Blazor\Sienar.Utils.Blazor.csproj", "{B737D83C-D052-44C9-B75D-4E0992AF953E}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "identity", "identity", "{D9A73344-DB2F-4065-B6E9-EEBE0601436F}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{1C5B004B-3A32-4D94-9583-568914BBE946}"
@@ -94,10 +92,6 @@ Global
 		{757F285A-17D6-4387-886B-FC44445F1C1B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{757F285A-17D6-4387-886B-FC44445F1C1B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{757F285A-17D6-4387-886B-FC44445F1C1B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B737D83C-D052-44C9-B75D-4E0992AF953E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B737D83C-D052-44C9-B75D-4E0992AF953E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B737D83C-D052-44C9-B75D-4E0992AF953E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B737D83C-D052-44C9-B75D-4E0992AF953E}.Release|Any CPU.Build.0 = Release|Any CPU
 		{BBF508D2-EDF5-4D6C-AEBE-EE4FAFB1F642}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{BBF508D2-EDF5-4D6C-AEBE-EE4FAFB1F642}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BBF508D2-EDF5-4D6C-AEBE-EE4FAFB1F642}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -132,7 +126,6 @@ Global
 		{FB99F387-A199-4CC3-9654-61ADD3087980} = {11C863C2-C181-4284-974F-9DEC271D6EE3}
 		{2FD53AC3-8BF0-4DD4-8DE8-AE0F61C077A3} = {AB411524-8DBB-4BC9-BD47-BA8822402E93}
 		{757F285A-17D6-4387-886B-FC44445F1C1B} = {2FD53AC3-8BF0-4DD4-8DE8-AE0F61C077A3}
-		{B737D83C-D052-44C9-B75D-4E0992AF953E} = {11C863C2-C181-4284-974F-9DEC271D6EE3}
 		{D9A73344-DB2F-4065-B6E9-EEBE0601436F} = {01DBB80A-3191-4D30-93E5-350E1C46E70C}
 		{1C5B004B-3A32-4D94-9583-568914BBE946} = {D9A73344-DB2F-4065-B6E9-EEBE0601436F}
 		{BBF508D2-EDF5-4D6C-AEBE-EE4FAFB1F642} = {1C5B004B-3A32-4D94-9583-568914BBE946}
