+++ plugins/core/src/Sienar.Plugins.Core.Client/Sienar.Plugins.Core.Client.csproj
@@ -36,12 +36,6 @@
 		<PackageReference
 			Condition=" '$(Configuration)' == 'Release' "
 			Include="Sienar.Utils" Version="0.1.1"/>
-		<ProjectReference
-			Condition=" '$(Configuration)' != 'Release' "
-			Include="..\..\..\..\shared\src\Sienar.Utils.Blazor\Sienar.Utils.Blazor.csproj"/>
-		<PackageReference
-			Condition=" '$(Configuration)' == 'Release' "
-			Include="Sienar.Utils.Blazor" Version="0.1.0"/>
 	</ItemGroup>
 
 </Project>
