+++ plugins/identity/src/Sienar.Plugins.Identity.Client/Sienar.Plugins.Identity.Client.csproj
@@ -46,10 +46,6 @@
 		<PackageReference
 			Condition=" '$(Configuration)' == 'Release' "
 			Include="Sienar.Ui" Version="0.1.0"/>
-		<ProjectReference Condition=" '$(Configuration)' != 'Release' " Include="..\..\..\..\shared\src\Sienar.Utils.Blazor\Sienar.Utils.Blazor.csproj" />
-		<PackageReference
-			Condition=" '$(Configuration)' == 'Release' "
-			Include="Sienar.Utils.Blazor" Version="0.1.0"/>
 	</ItemGroup>
 
 </Project>
