+++ shared/src/Sienar.Ui/Sienar.Ui.csproj
@@ -38,12 +38,6 @@
 		<PackageReference
 			Condition=" '$(Configuration)' == 'Release' "
 			Include="Sienar.Utils" Version="0.1.1"/>
-		<ProjectReference
-			Condition=" '$(Configuration)' != 'Release' "
-			Include="..\Sienar.Utils.Blazor\Sienar.Utils.Blazor.csproj"/>
-		<PackageReference
-			Condition=" '$(Configuration)' == 'Release' "
-			Include="Sienar.Utils.Blazor" Version="0.1.0"/>
 	</ItemGroup>
 
 </Project>
