--- shared/src/Sienar.Utils.Blazor/GlobalUsings.cs
@@ -1,23 +1,0 @@
-﻿global using Microsoft.AspNetCore.Components;
-global using Microsoft.AspNetCore.Components.Authorization;
-global using Microsoft.Extensions.Configuration;
-global using Microsoft.Extensions.DependencyInjection;
-global using Microsoft.Extensions.DependencyInjection.Extensions;
-global using Microsoft.Extensions.Logging;
-global using Microsoft.Extensions.Options;
-global using Microsoft.JSInterop;
-global using System;
-global using System.Collections.Generic;
-global using System.Diagnostics.CodeAnalysis;
-global using System.Linq;
-global using System.Threading.Tasks;
-global using Sienar.Configuration;
-global using Sienar.Data;
-global using Sienar.Extensions;
-global using Sienar.Hooks;
-global using Sienar.Html;
-global using Sienar.Infrastructure;
-global using Sienar.Menus;
-global using Sienar.Plugins;
-global using Sienar.Security;
-global using Sienar.Ui;
