--- shared/src/Sienar.Utils.Blazor/Sienar.Utils.Blazor.csproj
@@ -1,38 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Razor">
-
-	<PropertyGroup>
-		<TargetFramework>net8.0</TargetFramework>
-		<Nullable>enable</Nullable>
-		<RootNamespace>Sienar</RootNamespace>
-		<GenerateDocumentationFile>true</GenerateDocumentationFile>
-
-		<PackageId>Sienar.Utils.Blazor</PackageId>
-		<Title>Sienar.Utils.Blazor</Title>
-		<PackageVersion>0.1.0</PackageVersion>
-		<PackageReleaseNotes>Initial release</PackageReleaseNotes>
-		<Authors>Christian LeVesque</Authors>
-		<Description>Contains a library of Blazor utilities for use when authoring Sienar applications</Description>
-		<Copyright>Copyright 2024 Christian LeVesque</Copyright>
-		<PackageProjectUrl>https://sienar.io</PackageProjectUrl>
-		<PackageReadmeFile>README.md</PackageReadmeFile>
-		<RepositoryUrl>https://github.com/christianlevesque/sienar</RepositoryUrl>
-		<RepositoryType>git</RepositoryType>
-		<PackageTags>sienar framework utilities blazor plugin</PackageTags>
-		<PackageLicenseExpression>Apache-2.0</PackageLicenseExpression>
-	</PropertyGroup>
-
-	<ItemGroup>
-		<PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly" Version="8.0.0"/>
-		<PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly.Authentication" Version="8.0.0"/>
-	</ItemGroup>
-
-	<ItemGroup>
-		<ProjectReference
-			Condition=" '$(Configuration)' != 'Release' "
-			Include="..\Sienar.Utils\Sienar.Utils.csproj"/>
-		<PackageReference
-			Condition=" '$(Configuration)' == 'Release' "
-			Include="Sienar.Utils" Version="0.1.1"/>
-	</ItemGroup>
-
-</Project>
