+++ shared/src/Sienar.Utils/Extensions/EnumExtensions.cs
@@ -1,4 +1,5 @@
-﻿using System.Reflection;
+﻿using System.ComponentModel;
+using System.Reflection;
 
 namespace Sienar.Extensions;
