+++ shared/src/Sienar.Utils/Extensions/SienarUtilsBlazorJsRuntimeExtensions.cs
@@ -1,7 +1,7 @@
 ﻿namespace Sienar.Extensions;
 
 /// <summary>
-/// Contains <see cref="IJSRuntime"/> extension methods for the <c>Sienar.Utils.Blazor</c> assembly
+/// Contains <see cref="IJSRuntime"/> extension methods for the <c>Sienar.Utils</c> assembly
 /// </summary>
 public static class SienarUtilsBlazorJsRuntimeExtensions
 {
