+++ shared/src/Sienar.Utils/GlobalUsings.cs
@@ -1,11 +1,13 @@
-﻿global using Microsoft.Extensions.Configuration;
+﻿global using Microsoft.AspNetCore.Components;
+global using Microsoft.AspNetCore.Components.Authorization;
+global using Microsoft.Extensions.Configuration;
 global using Microsoft.Extensions.DependencyInjection;
 global using Microsoft.Extensions.DependencyInjection.Extensions;
 global using Microsoft.Extensions.Logging;
 global using Microsoft.Extensions.Options;
+global using Microsoft.JSInterop;
 global using System;
 global using System.Collections.Generic;
-global using System.ComponentModel;
 global using System.Diagnostics.CodeAnalysis;
 global using System.Linq;
 global using System.Threading.Tasks;
@@ -18,3 +20,4 @@ global using Sienar.Infrastructure;
 global using Sienar.Menus;
 global using Sienar.Plugins;
 global using Sienar.Security;
+global using Sienar.Ui;
