+++ shared/src/Sienar.Utils/Infrastructure/DefaultDelegateHandler.cs
