+++ shared/src/Sienar.Utils/Infrastructure/GlobalComponentProvider.cs
