+++ shared/src/Sienar.Utils/Infrastructure/IDelegateHandler.cs
