+++ shared/src/Sienar.Utils/Infrastructure/IScheduler.cs
