+++ shared/src/Sienar.Utils/Infrastructure/LayoutProvider.cs
