+++ shared/src/Sienar.Utils/Infrastructure/StateProviderBase.cs
