+++ shared/src/Sienar.Utils/Sienar.Utils.csproj
@@ -26,11 +26,23 @@
 	</ItemGroup>
 
 	<ItemGroup>
+		<PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly" Version="8.0.0"/>
+		<PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly.Authentication" Version="8.0.0"/>
 		<PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="8.0.2"/>
 		<PackageReference Include="Microsoft.Extensions.Configuration" Version="8.0.0"/>
 		<PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="8.0.0"/>
 		<PackageReference Include="Microsoft.Extensions.Options" Version="8.0.0" />
 		<PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="8.0.0"/>
+	</ItemGroup>
+
+	<ItemGroup>
+	  <AdditionalFiles Include="Ui\OptionalComponent.razor" />
+	  <AdditionalFiles Include="Ui\SienarApp.razor" />
+	  <AdditionalFiles Include="Ui\SienarFound.razor" />
+	  <AdditionalFiles Include="Ui\SienarHead.razor" />
+	  <AdditionalFiles Include="Ui\SienarRoutes.razor" />
+	  <AdditionalFiles Include="Ui\Username.razor" />
+	  <AdditionalFiles Include="Ui\_Imports.razor" />
 	</ItemGroup>
 
 </Project>
